+++ instances/parcel.cpp
@@ -30,41 +30,58 @@ namespace fs = std::filesystem;
 int main(int argc, const char** argv)
 {
   if (argc < 3) {
-    cout << "parcel <filename> [-pack|-unpack|-verify|-clean|-unzip|-merge|-help] " << endl;
     if (argc > 1) {
       std::string option = argv[1];
-      if (option == "-help")
+      if (option == "-help") {
         arc::Parcel("ignore").help();
+        return 0;
+      }
     }
-    return 0;
+    else {
+      cout << "parcel <filename> [-pack|-unpack|-verify|-clean|-unzip|-merge|-help] " << endl;
+      return -1;
+    }
   }
   try {
-    std::string option = argv[2];
+    std::string option = argc == 2 ? "-pack" : argv[2];
     auto filename = argv[1];
     FileNotFoundException::assertion(filename, __INFO__);
     Parameter parameter = ~extras::Paths(filename);
     arc::ParcelCmdLine parcel(parameter);
-    if (option == "-pack")
+    if (option == "-pack") {
       parcel.pack();
-    if (option == "-unpack")
+    }
+    if (option == "-unpack") {
       parcel.unpack();
-    if (option == "-verify")
+    }
+    if (option == "-verify") {
       parcel.verify_integrity();
-    if (option == "-clean")
+    }
+    if (option == "-clean") {
       parcel.clean();
-    if (option == "-cat")
+    }
+    if (option == "-cat") {
       parcel.cat();
-    if (option == "-merge")
+    }
+    if (option == "-merge") {
       parcel.merge();
-    if (option == "-dir")
+    }
+    if (option == "-dir") {
       parcel.dir();
-    if (option == "-unzip")
+    }
+    if (option == "-unzip") {
       parcel.unzip();
-    if (option == "-help")
+    }
+    if (option == "-help") {
       parcel.help();
-    return 0;
+    }
+    throw arc::UnknownOptionException(option, __INFO__);
   }
-  catch (exception& ex) {
+  catch (extras::exception& ex) {
+    cout << ex << endl;
+    return -1;
+  }
+  catch (std::exception& ex) {
     cout << ex.what() << endl;
     return -1;
   }