+++ changelog.md
@@ -4,8 +4,8 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), 
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).</br>
 > To add a tag to the git repository do:
-> > git tag -a v1.2.0 -m "Added directive"</br>
-> > git push origin v1.2.0
+> > git tag -a v1.3.0 -m "Added parcel & implode utilities"</br>
+> > git push origin v1.3.0
 > 
 
 IMPORTANT: 
@@ -28,6 +28,11 @@ Correct URL for github, (for source compare)</br>
 Adding mock test cases </br>
 Adding unit test cases </br>
 Updated README.md accordinging</br>
+
+#### 1.3.0 (2021-12-01)
+### Added 
+- Added parcel & implode utilities
+- Added HOWTO-...md files 
 
 #### 1.2.0 (2021-12-01)
 ### Added 
@@ -178,7 +183,8 @@ Updated README.md accordinging</br>
 ### Removed
 - Conan package manager deprecated
 
-[Unreleased]: https://github.com/perriera/extras_arc/compare/v1.2.0...HEAD
+[Unreleased]: https://github.com/perriera/extras_arc/compare/v1.3.0...HEAD
+[1.3.0]: https://github.com/perriera/extras_arc/compare/v1.2.0...v1.3.0
 [1.2.0]: https://github.com/perriera/extras_arc/compare/v1.1.2...v1.2.0
 [1.1.3]: https://github.com/perriera/extras_arc/compare/v1.1.2...v1.1.3
 [1.1.2]: https://github.com/perriera/extras_arc/compare/v1.1.1...v1.1.2